Adds new Quick Classification Method.
This method uses an indexed search to improve the performance of classification for a small hit during training. This method uses the indexed bins to determine where it is most likely to find similar samples in the training set. Then new test samples are only compared against items in that nearest bin. This results in far less overall comparisons and dramatically better classificaiton performance at the cost of some potential accuracy in the results.

More testing is currently needed to quantify the expected losses.